requirements.txt
apns2==0.7.1
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
DateTime==4.3
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.5.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-fsm==2.7.0
django-getpaid==2.2.1
django-getpaid-epay==0.1.4
django-heroku==0.3.1
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==4.0.0
django-push-notifications==2.0.0
django-redis==4.12.1
django-registration==3.1
django-rest-swagger==2.2.0
django-smart-selects==1.5.6
django-unixdatetimefield==1.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
fcm-django==0.3.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kkb==0.6
M2Crypto==0.36.0
MarkupSafe==1.1.1
numpy==1.18.5
openapi-codec==1.3.2
packaging==20.4
pendulum==2.1.2
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
py-vapid==1.7.1
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2020.1
pywebpush==1.11.0
redis==3.5.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
xmltodict==0.12.0
zope.interface==5.1.0
